backend/pyproject.toml
[project]
name = "resumate-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.3.4",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
    "email-validator>=2.3.0",
    "google-auth>=2.28.1",
    "httpx>=0.27.0",
    "requests>=2.31.0",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
]

[dependency-groups]
dev = []
